backend/requirements.txt
fastapi==0.115.6
uvicorn==0.24.0
mysql-connector-python==9.1.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1